[IMP] runbot: exclude l10n_ modules unless explicit
As of Odoo 9 the first l10n module installed will automatically
install its chart of account on the main company, so installing
the first one (l10n_ae) is surprising/strange.

Branches that want to build l10n_* modules should list
them explicitly in the "modules to install" field (modules).

In addition, having demo data to set the country of the
main company will automatically installe the relevant
l10n module if it exists. This is the default in Odoo 9
with l10n_us.