[REF] runbot: make jobs locking work in python3
Since PEP-446 file descriptors are non-inheritable by default.
A new method (os.set_inheritable) was introduced to explicitely make
fd inheritable. Also, the close_fds parameter of the subprocess.Popen
method is now True by default.
Finally, PEP-3151 changed the execpetion raised by fcntl.flock was
changed from IOError to OSError (and IOError became an alias of
OSError).

As a consequence of all that, the runbot locking mechanism to check if a
job is finished was not working in python3. Hence this commit.